django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "django-ninja>=1.0",
    "passlib[bcrypt]>=1.7",
    "psycopg2-binary>=2.9",
    "pyjwt[crypto]>=2.8",
    "python-dotenv>=1.0",
]
